* add base pages & routing:   ◊ Home   ◊ Settings   ◊ Detail * if click on PizaImage - go to Detail * send params from Home to Detail * move Search & Modal to Header component
