+++ screens/Home.jsx
@@ -1,0 +1,246 @@
+import { useEffect, useState } from 'react';
+import { useNavigation, useRoute } from '@react-navigation/native';
+import {
+  StyleSheet,
+  SafeAreaView,
+  View,
+  Text,
+  FlatList,
+  TouchableOpacity,
+  RefreshControl,
+} from 'react-native';
+import { itemsTemplate, generateItems } from '../common/templates/item-card';
+
+import { LinearGradient } from 'expo-linear-gradient';
+import CustomHeader from '../components/CustomHeader';
+import ItemCard from '../components/productViews/ItemCard';
+import { colors } from '../common/colors/colors';
+
+let DEBUG_MENU = true;
+
+export default function HomeScreen(props) {
+  const [loading, setLoading] = useState(true);
+  const [nextItem, setNextItem] = useState({ count: 0, start: 0 });
+  const [items, setItems] = useState([]);
+  const [filteredItems, setFilteredItems] = useState(items);
+  const [refreshing, setRefreshing] = useState(false);
+  const [freezeUpdate, onFreezeUpdate] = useState(false);
+  const [detailData, setDetailData] = useState({});
+
+  const navigation = useNavigation();
+  const route = useRoute();
+
+  const onDetailPress = () => {
+    navigation.navigate('Settings', {
+      ...detailData,
+      test: 'From Pizza Page',
+    });
+  };
+
+  let isItemsLoaded = false;
+  useEffect(() => {
+    if (!isItemsLoaded) {
+      const firstGettedItems = generateItems(4, 0); // get first 5 pizzas from server :)
+      setItems(firstGettedItems);
+      setFilteredItems(firstGettedItems);
+      setNextItem({ count: 1, start: 4 });
+      isItemsLoaded = true;
+    }
+  }, []);
+
+  const onRefresh = () => {
+    setRefreshing(true);
+    const nextGettedItems = generateItems(nextItem.count, nextItem.start);
+    setTimeout(
+      (nextGettedItems) => {
+        setItems((prev) => [...nextGettedItems, ...prev]);
+        setNextItem((prev) => ({
+          count: 1,
+          start: prev.start + 1,
+        }));
+        setRefreshing(false);
+      },
+      3000,
+      nextGettedItems
+    );
+  };
+
+  const onUpdateEndList = ({ distanceFromEnd }) => {
+    if (!refreshing && !freezeUpdate && nextItem.start < 15) {
+      // max 40
+      const nextGettedItems = generateItems(5, nextItem.start);
+      setTimeout(
+        (nextGettedItems) => {
+          setNextItem((prev) => ({ ...prev, start: prev.start + 5 }));
+          setItems((prev) => [...prev, ...nextGettedItems]);
+          setRefreshing(false);
+        },
+        500,
+        nextGettedItems
+      );
+    }
+    // else if (!freezeUpdate && nextItem.start > 14)
+    //   Alert.alert('Sorry, End Pizza today.');
+  };
+
+  // add new getted from server pizzas to list
+  useEffect(() => {
+    setFilteredItems(items);
+  }, [items]);
+
+  // simulate loader
+  useEffect(() => {
+    if (isItemsLoaded) {
+      setTimeout(() => setLoading(false), 2000);
+    }
+  }, []);
+
+  const onSearch = (e) => {
+    /**
+     *  TODO add search type - price, description, favorites, promo
+     */
+    const filtered = items.filter((item) => {
+      return item.title.toLowerCase().includes(e.toLowerCase());
+    });
+    setFilteredItems(filtered);
+  };
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <LinearGradient
+        colors={colors['app-background-gradient']}
+        style={styles.container}
+      >
+        <CustomHeader
+          onSearch={onSearch}
+          onFreezeUpdate={onFreezeUpdate}
+          loading={loading}
+        />
+
+        <View style={styles.wrapper}>
+          <TouchableOpacity onPress={onDetailPress}>
+            <Text style={styles.warningText}>Home Screen</Text>
+          </TouchableOpacity>
+        </View>
+
+        <FlatList
+          style={[styles.listItems, { display: loading ? 'none' : 'flex' }]}
+          data={filteredItems}
+          // renderItem={({ item }) => (
+          //   <TouchableOpacity
+          //     style={styles.warningText}
+          //     onPress={() => navigation.navigate('Detail', { test: 'WORK!' })}
+          //   >
+          //     <Text
+          //       style={{
+          //         fontSize: 20,
+          //         color: '#333',
+          //         padding: 20,
+          //         width: '96%',
+          //         backgroundColor: colors['card-background'],
+          //         shadowColor: '#000',
+          //         shadowOffset: {
+          //           width: 0,
+          //           height: 7,
+          //         },
+          //         shadowOpacity: 0.41,
+          //         shadowRadius: 9.11,
+          //         elevation: 14,
+          //       }}
+          //     >
+          //       {item.title}
+          //     </Text>
+          //   </TouchableOpacity>
+          // )}
+          renderItem={({ item }) => (
+            <TouchableOpacity
+              style={styles.warningText}
+              // onPress={() => navigation.navigate('Detail', { test: 'WORK!' })}
+            >
+              <ItemCard
+                item={item}
+                onPress={() =>
+                  navigation.navigate('Detail', { test: 'WORK!' + item.title })
+                }
+              />
+            </TouchableOpacity>
+          )}
+          // keyExtractor={({ item, index }) => console.warn(item, index)}
+          onPress={(data) => console.warn('TEST onPress: ', data)}
+          ListEmptyComponent={
+            <Text style={styles.warningText}>There is nothing</Text>
+          }
+          refreshing={refreshing}
+          refreshControl={
+            <RefreshControl
+              refreshing={refreshing}
+              onRefresh={onRefresh}
+              progressViewOffset={0.15}
+            />
+          }
+          onEndReachedThreshold={0.15}
+          onEndReached={onUpdateEndList} // can CONFLICT with search bar - fixed
+        />
+      </LinearGradient>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  wrapper: {
+    flex: 0.1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  container: {
+    flex: 1,
+    width: '100%',
+    alignItems: 'center',
+    justifyContent: 'flex-start',
+    backgroundColor: colors['primary-dark'], //colors['app-background'],
+  },
+  loaderContainer: {
+    flex: 1,
+    width: '100%',
+    height: '100%',
+    justifyContent: 'center',
+  },
+  listContainer: {
+    flex: 1,
+    width: '100%',
+    backgroundColor: colors['card-background'],
+  },
+  HeaderIconsWrapper: {
+    // flex: 1,
+    justifyContent: 'flex-end',
+    alignItems: 'baseline',
+    width: '96%',
+    // marginTop: 40,
+    // marginBottom: 20,
+    flexDirection: 'row',
+  },
+  searchButton: {
+    flex: 1,
+    maxWidth: '90%',
+  },
+  openModalButton: {
+    flex: 1,
+    maxHeight: 200,
+    padding: 10,
+    color: colors['primary-light'],
+  },
+  listItems: {
+    flex: 1,
+    width: '100%',
+
+    // borderWidth: 3,
+    // borderColor: 'red',
+  },
+  warningText: {
+    textAlign: 'center',
+    fontSize: 20,
+    color: colors['warning-text'],
+    padding: 20,
+    marginTop: 20,
+  },
+});
